CCC_project/back_end/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
python-dotenv==1.2.1
requests_cache==1.2.1
retry-requests

numpy==2.3.3
pandas==2.3.3
scikit-learn==1.7.2
xgboost==3.1.1
tensorflow==2.20.0
joblib==1.3.2
matplotlib==3.10.0
seaborn==0.13.2
PuLP==3.3.0
requests==2.32.5
openmeteo_requests==1.7.3
openmeteo_sdk==1.21.2
